remove third window and unify functions trees
