+++ src/main/java/com/redhat/devtools/intellij/knative/actions/RefreshAction.java
@@ -18,6 +18,8 @@ import com.redhat.devtools.intellij.knative.tree.KnEventingNode;
 import com.redhat.devtools.intellij.knative.tree.KnFunctionNode;
 import com.redhat.devtools.intellij.knative.tree.KnFunctionsNode;
 import com.redhat.devtools.intellij.knative.tree.KnFunctionsTreeStructure;
+import com.redhat.devtools.intellij.knative.tree.KnLocalFunctionNode;
+import com.redhat.devtools.intellij.knative.tree.KnLocalFunctionsNode;
 import com.redhat.devtools.intellij.knative.tree.KnRevisionNode;
 import com.redhat.devtools.intellij.knative.tree.KnRootNode;
 import com.redhat.devtools.intellij.knative.tree.KnServiceNode;
@@ -29,7 +31,8 @@ import javax.swing.tree.TreePath;
 public class RefreshAction extends StructureTreeAction {
     public RefreshAction() {
         this(KnRootNode.class, KnServingNode.class, KnServiceNode.class, KnRevisionNode.class,
-                KnEventingNode.class, KnFunctionsNode.class, KnFunctionNode.class);
+                KnEventingNode.class, KnFunctionsNode.class, KnFunctionNode.class,
+                KnLocalFunctionsNode.class, KnLocalFunctionNode.class);
     }
 
     public RefreshAction(Class... filters) {
