+++ src/main/java/com/redhat/devtools/intellij/knative/tree/KnLocalFunctionsNode.java
@@ -1,0 +1,9 @@
+package com.redhat.devtools.intellij.knative.tree;
+
+import org.jetbrains.annotations.NotNull;
+
+public class KnLocalFunctionsNode extends ParentableNode<KnRootNode> {
+    protected KnLocalFunctionsNode(@NotNull KnRootNode rootNode, @NotNull KnRootNode parent) {
+        super(rootNode, parent, "Local Functions");
+    }
+}
